Move addComment to ducks/fetchSource
